# server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gem Auto Rentals API
  description: |
    RESTful API for Gem Auto Rentals - a car rental management platform.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Roles
    - **CUSTOMER**: Regular users who can book vehicles
    - **SUPPORT**: Staff who can view and manage bookings
    - **MANAGER**: Staff with additional fleet management access
    - **ADMIN**: Full system access

    ## Rate Limiting
    API requests are rate limited to 100 requests per 15 minutes per IP address.
  version: 1.0.0
  contact:
    name: Gem Auto Rentals Support
    email: support@gemautorentals.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.gemautorentals.com/api
    description: Production server
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Get paginated customer list (staff only)
      operationId: listCustomers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by email, first name, or last name
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Customers retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer details
      description: Get customer profile (staff or own profile)
      operationId: getCustomer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/CustomerDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Customers]
      summary: Update customer
      description: Update customer information (staff or own profile)
      operationId: updateCustomer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Customers]
      summary: Delete customer
      description: Delete a customer account (admin only, cannot delete with active bookings)
      operationId: deleteCustomer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Cannot delete customer with active bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /customers/{id}/profile:
    put:
      tags: [Customers]
      summary: Update own profile
      description: Update your own profile including password
      operationId: updateProfile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Customer'
                  message:
                    type: string
        '400':
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /customers/{id}/bookings:
    get:
      tags: [Customers]
      summary: Get customer bookings
      description: Get all bookings for a customer
      operationId: getCustomerBookings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Customer bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== Payments ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
...
    CustomerId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Customer/User ID (CUID)
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not found
  schemas:
...
    VehicleCategory:
      type: string
      enum: [ECONOMY, STANDARD, PREMIUM, LUXURY, SUV, VAN]
...
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, ACTIVE, COMPLETED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [PENDING, PROCESSING, SUCCEEDED, FAILED, REFUNDED, PARTIALLY_REFUNDED]
    UserRole:
      type: string
      enum: [CUSTOMER, SUPPORT, MANAGER, ADMIN]

    # ===== Common Schemas =====
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    BookingExtras:
      type: object
      properties:
        insurance:
          type: boolean
        gps:
          type: boolean
        childSeat:
          type: boolean
        additionalDriver:
          type: boolean
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        vehicleId:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/BookingStatus'
        pickupLocation:
          type: string
        dropoffLocation:
          type: string
        dailyRate:
          type: number
        totalAmount:
          type: number
        extras:
          $ref: '#/components/schemas/BookingExtras'
        notes:
          type: string
          nullable: true
        vehicle:
          type: object
          properties:
            id:
              type: string
            make:
              type: string
            model:
              type: string
            year:
              type: integer
            images:
              type: array
              items:
                type: string
            category:
              $ref: '#/components/schemas/VehicleCategory'
        user:
          type: object
          properties:
            id:
              type: string
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
        payment:
          type: object
          nullable: true
          properties:
            id:
              type: string
            status:
              $ref: '#/components/schemas/PaymentStatus'
            amount:
              type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BookingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Booking'
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer

    # ===== Customer Schemas =====
    Customer:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        emailVerified:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            bookings:
              type: integer
            documents:
              type: integer
    CustomerDetail:
      allOf:
        - $ref: '#/components/schemas/Customer'
        - type: object
          properties:
            bookings:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    $ref: '#/components/schemas/BookingStatus'
                  startDate:
                    type: string
                    format: date-time
                  endDate:
                    type: string
                    format: date-time
                  totalAmount:
                    type: number
                  vehicle:
                    type: object
                    properties:
                      id:
                        type: string
                      make:
                        type: string
                      model:
                        type: string
                      year:
                        type: integer
            documents:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  type:
                    type: string
                  verified:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
            _count:
              type: object
              properties:
                bookings:
                  type: integer
                documents:
                  type: integer
                reviews:
                  type: integer
    UpdateCustomerRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        phone:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
          description: Admin only
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        phone:
          type: string
        currentPassword:
          type: string
          description: Required when changing password
        newPassword:
          type: string
          minLength: 8
    CustomerListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer

    # ===== Payment Schemas =====
...